Step 2: Install Webpack
Runs `npm install --save-dev webpack` to install webpack.

The `--save-dev` option writes the dependency to `package.json` so the
dependency can be tracked and installed on other systems by simply running
`npm install`. We use this option rather than `--save` because the
dependency is not needed in the production environment.

The `webpack` package is installed in `node_modules`. This directory
is added to `.gitignore`.