Step 5: Setup Multiple Entry Points and Common Chunk
We add a few page scripts that import different modules.

But we notice each entry point is getting the entire webpack runtime when
only `site.js` needs it. And we want any modules used by more than one
entry point to be loaded only once. So we configure `site.js` to be the
common chunk.