Step 7: Alias React as React Addons
In the last step we notice that both React and React with Addons are being
compiled into our scripts. This is bad.

To fix this we can alias the module `react` as `react/addons` so that any
time we require React we get React with Addons, resulting in the same
module being required either way. This allows webpack to identify the React
module and place it in the common chunk instead of loading both version.

If you aren't using React Addons, only require React, this step is not necessary.