Step 9: Configure JSX Loader
In the last step we see that webpack doesn't really know what to do with a
`.jsx` file. We need to tell it how to handle `.jsx` file by using a
"loader".

Webpack uses loaders to figure out how to process a file based on a file
extension pattern.

We install `jsx-loader` via npm. Then we list the loader in our config.