Step 11: Process Both .js and .jsx as JSX
It is tedious to remember to put the .jsx extension on any file that contains even the
smallest amount of jsx, while the rest of the file may be vanilla.

We can handle that problem by processing js and jsx files with the jsx
loader. To do that we just modify the loader test pattern.

NOTE: This might not be a best practice. But it shows the flexibility of
webpack.